同步全量实验版本：更新代码、配置、文档与 fig_full 结果图

+++ SPANN-PostFiltering/src/config.h
@@ -45,7 +45,7 @@ struct SPANNConfig {
     // ── PostFiltering parameters ──
     size_t k = 10;
     size_t num_warmup = 20;
-    size_t overfetch_factor = 50;     // fixed overfetch: k' = k * factor
+    double overfetch_factor = 50;     // fixed overfetch: k' = k * factor (may be fractional)
     bool overfetch_adaptive = true;   // adaptive overfetch based on selectivity
     bool batch_query = false;
